xform: multiple MIN/MAX subqueries rewrite, filter support
Previously, the ReplaceScalarMinMaxWithScalarSubqueries rewrite
rule only rewrites queries with no WHERE clause, for example:
    select min(pk_col1), max(pk_col1) from t1;
becomes:
    select (select min(pk_col1) from t1),
           (select max(pk_col1) from t1);

... allowing each subquery to utilize a limited scan
    via the primary key.

This was inadequate because it is of limited value.
Most user queries have a WHERE clause, so WHERE clause
support is crucial to allow this optimization to be
applied in more cases.

To address this, this patch adds a new custom match function
named SimpleScanOrSelect for the ReplaceScalarMinMaxWithScalarSubqueries
rewrite to detect not just Scans, but Selects as input to
scalar group by expressions.  If a Select is present, a
new Select is built with Filters (WHERE clause) mapped to
new column IDs in a copy of the tree. The new Select is
added as the source of each subquery's manufactured
ScalarGroupByExpr.

Release note (performance improvement): A SELECT query with both
MIN(LeadingIndexColumn) and MAX(LeadingIndexColumn) and a
WHERE clause can now be performed with two LIMITED SCANs
instead of a single FULL SCAN, for example:
    create table t1 (a int, b int, primary key(a));
    explain select min(a), max(a) from mws.t1 where b < 6;