opt: Push projection into UNIONed branches of SplitDisjunction
Previously, when ORed predicates were split to explore a unioned index
access path plan, the projection was not pushed into each branch of
the UNION, and each branch was treated as if it needs to access all of
the columns in all predicates, leading to unnecessary index joins back
to the base table to get the missing column data.  For example, index
t1_ix2 has columns b and a (primary key columns are included in indexes)
and index t1_ix3 has columns c and a, yet the following uses index join:
```
CREATE TABLE t1(a int, b int, c int, primary key(a));
create index t1_ix2 on t1(b);
create index t1_ix3 on t1(c);
insert into t1 select g,g,g FROM generate_series(1,100) g(g);
ANALYZE T1;
EXPLAIN(opt) SELECT a FROM t1 WHERE b=1 or c=2;
-- The query is rewritten internally to:
-- SELECT DISTINCT a FROM (SELECT a FROM t1 where b=1 UNION ALL
--                         SELECT a FROM t1 where c=1);

                           info
----------------------------------------------------------
  project
   └── distinct-on
        ├── union-all
        │    ├── index-join t1
        │    │    └── scan t1@t1_idx2
        │    │         └── constraint: /8/7: [/1 - /1]
        │    └── index-join t1
        │         └── scan t1@t1_idx3
        │              └── constraint: /14/12: [/2 - /2]
        └── aggregations
             ├── const-agg
             │    └── b
             └── const-agg
                  └── c
```

This was inadequate because it can be costly to perform an index join,
especially when the base table rows reside on a different node than
the corresponding index rows. When index join is unnecessary it should
not be performed.

To address this, this patch records in a new SelectPrivate member of
SelectExpr the set of projected columns in order to allow the
SplitDisjunction and SplitDisjunctionAddKey rules to push the projection
into each branch of the UNION or UNION ALL operation. A different
approach was initially attempted which added new exploration rules which
matched on the ProjectExpr instead of the SelectExpr. Unfortunately this
caused an explosion in the search space for complex SQL expressions
since both the original transformation rule and the new one could match,
leading to unacceptibly long optimization time. Both rules would be
needed because we don't want to miss out on the SplitDisjunction
optimization when there is no projection.

Release note (performance improvement): Queries with ORed WHERE clause
predicates on different indexes previously performed unnecessary index
joins back to the base table, causing them to run slower. To fix this,
we pushed the projection (set of selected columns) into each branch of
the optimizer-generated UNION or UNION ALL operation, so that
unnecessary columns which may trigger an index join are not accessed.