tests: fix fatal errors from timeout-out statements in queryComparisonTests
Fixes #95680

Roachtests such as tlp, costfuzz or unoptimized-query-oracle may time
out helper statements such as `SET testing_optimizer_random_seed = xxx`
and misinterpret the error as a fatal error, causing an issue to be
opened.

This fixes the problem by not erroring out the test if the test or round
timeout has expired. An alternative fix could have been to avoid
throwing errors for SET statements, with the idea that such errors are
likely rare. The current fix was chosen so that we still might catch
such errors if they were to occur.

Release note: None