sql: fast path uniqueness checks for single-row insert
This adds support for building and executing simple INSERT statements
with a single-row VALUES clause where any required uniqueness constraint
checks can be handled via a constrained scan on an index.

This includes INSERT cases such as:

- a single-row VALUES clause into a REGIONAL BY ROW table with a
PRIMARY KEY which has a UUID column generated by default, ie.
`id UUID PRIMARY KEY DEFAULT gen_random_uuid()`, where the
crdb_region column is not specified in the VALUES clause; either
the gateway region is used or it is computed based on other column
values.
- a single-row VALUES clause into a REGIONAL BY ROW table with a
hash sharded unique index where the crdb_region column is not specified
in the VALUES clause
- a single-row VALUES clause into a table with an experimental UNIQUE
WITHOUT INDEX constraint which also has an index with leading key
columns matching the constraint columns.

In optbuild, when creating a uniqueness check for rows which are added
to a table, a fast path index check relation is also built when the
mutation source is a single-row values expression or WithScan from
a single-row values expression. That relation is a filtered
Select of a Scan from the target table, where the filters equate
all of the unique check columns with their corresponding
constants or placeholders from the Values expression. Any necessary
partial index predicates are also added to the filters. The relation
is added as element `FastPathIndexCheck` in `UniqueChecksItem`, and
is explored just like any other relation in the exploration phase.

A new exploration rule called InsertFastPath is added to walk the memo
group members created during exploration in the `FastPathIndexCheck`,
to find any which have been rewritten as a constrained `ScanExpr`. If
found, that means that Scan fully represents the lookup needed to check
for duplicate entries and the Scan constraint can be used to identify
the constants to use in a KV lookup on the scanned index in a fast
path check.

Function CanUseUniqueChecksForInsertFastPath walks the expressions
generated during exploration of the `FastPathIndexCheck` relation.
If a constrained scan is found, it is used to build a
`InsertFastPathFKUniqCheck` structure, shared by foreign key fast
path checking, to communicate to the execbuilder the table and index
to use for the check, and the column ordinals in the insert row
to use for building the fast path KV request. In addition, a new
`DatumsFromConstraint` field is added, which allows an index key
column to be matched with multiple Datums, so one insert row may result
in more than one KV lookup for a given uniqueness constraint. The new
`InsertFastPathFKUniqCheck` structures are place in the private sections
of the corresponding `UniqueChecksItem`s and communicated to the caller
via return value `newUniqueChecks`. If the input to the insert statement
is a projection on top of a Values expression, it is merged, if
possible, with the Values expression to enable more fast path cases.
This expression is returned to the caller via `insInput`.

A small adjustment is made in the coster to make the fast path unique
constraint slightly cheaper, so it should always be chosen over the
original non-fast path check.

Epic: CRDB-26290
Fixes: #58047

Release note (performance improvement): This patch adds support for
insert fast-path uniqueness checks on REGIONAL BY ROW tables where
the source is a VALUES clause with a single row. This results a
reduction in latency for single-row inserts to REGIONAL BY ROW tables
or hash-sharded REGIONAL BY ROW tables with unique indexes.