builtins: avoid overflow in sum square difference calculation
For very large floating point values, the var_pop aggregate function may
easily overflow when calculating the intermediate result of the sum square
difference calculation on the very first row processed in the final aggregation
step of a distributed query.

The calculation is:
```
a.sqrDiff += sqrDiff + delta*delta*float64(count)*float64(a.count)/float64(totalCount)
```

As there have been no results computed yet, the current mean value is
zero, and delta, the difference between the input and the mean, is the
input value itself. When this is a large value, greater than
1.340781e+154, the calculation `delta*delta` yields `+Inf` and the
result remains `+Inf` even though the value is later multiplied by a
count of zero (the running count of number of values).

The solution is to change the order of operations to multiply the counts
first, and divide delta directly by totalCount, to avoid overflow in as
many cases as possible.

Note, this fixes the single-row case, but does not fix cases where the
calculation of the sum, which is an input to `floatSumSqrDiffsAggregate.Add`,
has overflowed due to summing many large values.

Epic: none
Fixes: #109334

Release note: None